+++ ClangPowerTools/ClangPowerTools/Commands/CommandsController.cs
@@ -2,6 +2,7 @@
 using EnvDTE80;
 using Microsoft.VisualStudio.Shell;
 using System;
+using ClangPowerTools.Handlers;
 
 namespace ClangPowerTools
 {
@@ -9,14 +10,14 @@ namespace ClangPowerTools
   {
     #region Members
 
-    private IServiceProvider mServiceProvider;
+    private IAsyncServiceProvider mServiceProvider;
     private DTE2 mDte;
 
     #endregion
 
     #region Constructor
 
-    public CommandsController(IServiceProvider aServiceProvider, DTE2 aDte)
+    public CommandsController(IAsyncServiceProvider aServiceProvider, DTE2 aDte)
     {
       mDte = aDte;
       mServiceProvider = aServiceProvider;
@@ -35,7 +36,7 @@ namespace ClangPowerTools
 
     public void AfterExecute()
     {
-      DispatcherHandler.Invoke(() =>
+      UIUpdater.Invoke(() =>
       {
         Running = false;
       });
@@ -43,7 +44,7 @@ namespace ClangPowerTools
 
     public void QueryCommandHandler(object sender, EventArgs e)
     {
-      DispatcherHandler.Invoke(() =>
+      UIUpdater.Invoke(() =>
       {
         //var itemsCollector = new ItemsCollector(mPackage);
         //itemsCollector.CollectSelectedFiles(mDte, ActiveWindowProperties.GetProjectItemOfActiveWindow(mDte));
@@ -69,7 +70,7 @@ namespace ClangPowerTools
           command.Visible = true;
         }
       });
-     
+
     }
 
     public void OnBuildBegin(vsBuildScope Scope, vsBuildAction Action)