+++ ClangPowerTools/ClangPowerTools/Commands/SettingsCommand.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.ComponentModel.Design;
+using EnvDTE80;
 using Microsoft.VisualStudio.Shell;
 
 namespace ClangPowerTools
@@ -16,9 +17,12 @@ namespace ClangPowerTools
     /// Adds our command handlers for menu (commands must exist in the command table file)
     /// </summary>
     /// <param name="package">Owner package, not null.</param>
-    public SettingsCommand(Package aPackage, Guid aGuid, int aId) : base(aPackage, aGuid, aId)
+    public SettingsCommand(DTE2 aDte, AsyncPackage aPackage, Guid aGuid, int aId)
+      : base(aDte, aPackage, aGuid, aId)
     {
-      if (ServiceProvider.GetService(typeof(IMenuCommandService)) is OleMenuCommandService commandService)
+      var commandService = ServiceProvider.GetService(typeof(IMenuCommandService)) as OleMenuCommandService;
+
+      if (null != commandService)
       {
         var menuCommandID = new CommandID(CommandSet, Id);
         var menuItem = new MenuCommand(this.ShowSettings, menuCommandID);
@@ -28,7 +32,8 @@ namespace ClangPowerTools
 
     #endregion
 
-    #region Command
+
+    #region Private Methods
 
     /// <summary>
     /// This function is the callback used to execute the command when the menu item is clicked.
@@ -39,7 +44,7 @@ namespace ClangPowerTools
     /// <param name="e">Event args.</param>
     private void ShowSettings(object sender, EventArgs e)
     {
-      Package.ShowOptionPage(typeof(ClangGeneralOptionsView));
+      AsyncPackage.ShowOptionPage(typeof(ClangGeneralOptionsView));
     }
 
     #endregion