+++ ClangPowerTools/ClangPowerTools/Helpers/DispatcherHandler.cs
@@ -1,6 +1,6 @@
 ﻿using EnvDTE80;
 using System;
-using System.Windows.Interop;
+using System.Windows;
 using System.Windows.Threading;
 
 namespace ClangPowerTools
@@ -18,20 +18,20 @@ namespace ClangPowerTools
 
     public static void Initialize(DTE2 aDte)
     {
-      mDispatcher = HwndSource.FromHwnd((IntPtr)aDte.MainWindow.HWnd).RootVisual.Dispatcher;
+      mDispatcher = Application.Current.Dispatcher;
     }
 
-    public static void BeginInvoke(Action aAction)
+    public static void BeginInvoke(Action aAction, DispatcherPriority aPrioriry)
     {
-      mDispatcher.BeginInvoke(DispatcherPriority.Normal, new Action(() =>
+      mDispatcher.BeginInvoke(aPrioriry, new Action(() =>
       {
-        aAction.Invoke();
+        aAction.BeginInvoke(aAction.EndInvoke, null);
       }));
     }
 
-    public static void Invoke(Action aAction)
+    public static void Invoke(Action aAction, DispatcherPriority aPrioriry)
     {
-      mDispatcher.Invoke(DispatcherPriority.Normal, new Action(() =>
+      mDispatcher.Invoke(aPrioriry, new Action(() =>
       {
         aAction.Invoke();
       }));