+++ ClangPowerTools/ClangPowerTools/Output/OutputManager.cs
@@ -2,12 +2,15 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Text;
+using System.Windows.Threading;
 using EnvDTE80;
+using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Shell.Interop;
+using ClangPowerTools.Handlers;
 
 namespace ClangPowerTools
 {
-    public class OutputManager
+  public class OutputManager
   {
     #region Members
 
@@ -57,7 +60,7 @@ namespace ClangPowerTools
 
     public void Clear()
     {
-      DispatcherHandler.BeginInvoke(() =>
+      UIUpdater.Invoke(() =>
       {
         mOutputWindow.Clear();
       });
@@ -65,7 +68,7 @@ namespace ClangPowerTools
 
     public void Show()
     {
-      DispatcherHandler.BeginInvoke(() =>
+      UIUpdater.Invoke(() =>
       {
         mOutputWindow.Show(mDte);
       });
@@ -73,7 +76,7 @@ namespace ClangPowerTools
 
     public void AddMessage(string aMessage)
     {
-      DispatcherHandler.BeginInvoke(() =>
+      UIUpdater.Invoke(() =>
       {
         if (String.IsNullOrWhiteSpace(aMessage))
           return;